backend/requirements.txt
fastapi
uvicorn[standard]
pytest
databricks-sql-connector
